Rework how the profiles aare stored and handled for dev and prod builds
The app.json file is not intended to be stored in the relese build. It's
only a definition files for building the app. Only one key in this JSON
is propagated in the build. It's the xcraft key which contents the
settings associated to the modules. These settings are handled by
xcraft-core-etc.

With this commits, the profiles are moved in the
xcraft.goblin-configurator entry because it's owned by this module. The
fallback system is reproduced by supporting a new style of key names for
the settings. If the key is goblin-configurator, then it's the settings
for the module of this app. But if the key is goblin-configurator@appId
where appId is the id of an other app, then the settings are retrieved
in the other app.

The real advantage here it's that now possible to use the @appId trick
with all sort of settings and not only with the profiles because it's
generic.