Rework open-session (frontend) by passing the profile key or the feed
We use two different arguments for the readability. The number argument
is no longer used.

The profile key can be retrieved when only the feed is passed, because
the desktop saves this key in its state. Then it's possible to provide
the full configuration payload when a session is already opened.